fix margin elements
