--- app/a-propos/loading.tsx
@@ -1,48 +1,0 @@
-// import { Header } from '@/components';
-
-export default function Loading() {
-  return (
-    <div className=""></div>
-    // <main className="w-full">
-    //   <div className="px-[22px] md:px-16 lg:px-28 pt-5 pb-[35px] lg:pb-[53px] bg-primary animate-pulse">
-    //     <Header color="secondary" />
-    //     <div className="mt-[38px] lg:mt-[62px] text-f-secondary flex flex-wrap md:flex-nowrap gap-5 md:gap-[46px]">
-    //       <div className="w-[346px] h-[312px] md:w-[546px] md:h-[633px] relative rounded-[20px]"></div>
-    //       <div className="grow">
-    //         <h2 className="font-GothamLight font-medium text-[12px] lg:leading-[19.12px] lg:text-[18px] antialiased mb-2  bg-slate-700"></h2>
-    //         <h1 className="uppercase aqua-banner text-[20px] lg:text-[50px] leading-none antialiased mb-5 lg:mb-12 bg-slate-700"></h1>
-    //         <div className="mb-[31px] md:mb-12 flex flex-wrap gap-3 lg:max-w-none lg:gap-5 w-full bg-slate-700">
-    //           {/* {TAGS.map((tag, index) => (
-    //             <Button key={index} className="w-fit">
-    //               {tag}
-    //             </Button>
-    //           ))} */}
-    //         </div>
-    //         <div className="mb-[30px] aqua-caption text-[12px] lg:text-[16px] lg-leading-[22px] max-w-[640px] antialiased">
-    //           {/* La création a toujours été une composante essentielle de mon
-    //           existence. */}
-    //           <div className="w-full bg-slate-700" />
-    //           <br />
-    //           <div className="w-full bg-slate-700" />
-    //           {/* Mon intérêt pour l'art s'est manifesté à travers ma passion pour
-    //           la danse ainsi que mon attrait pour le dessin. C'est donc tout
-    //           naturellement que j'ai choisi de suivre une carrière créative. */}
-    //           <br />
-    //           <div className="w-full bg-slate-700" />
-    //           {/* Ce métier représente l'harmonie idéale entre les aspects
-    //           commerciaux et créatifs indispensables à mon épanouissement. */}
-    //         </div>
-    //         <div className="mb-[30px] lg:mb-[60px] bg-slate-700">
-    //           {/* <Button
-    //             variant="tertiary"
-    //             className="rounded-[5px] lg:rounded-[20px] lg:text-[25px] py-4 lg:py-[22px]" //h-7 lg:h-[71px]
-    //           >
-    //             Je veux te contacter ! ✨
-    //           </Button> */}
-    //         </div>
-    //       </div>
-    //     </div>
-    //   </div>
-    // </main>
-  );
-}
