--- app/contact/layout.tsx
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export default function ContactLayout({ children }: { children: ReactNode }) {
-  return <section>{children}</section>;
-}
